Arreglado rutas, Anadido rutas nuevas, creado una nueva ruta /,
