+++ backend/src/routes/user.routes.js
@@ -1,5 +1,6 @@
 const { Router } = require("express");
 const ctrlUser = require("../controller/user.controller");
+const util = require("../utils/utils")
 
 class UserRoutes {
   constructor() {
@@ -7,16 +8,7 @@ class UserRoutes {
     this.#routes();
   }
 
-  #session_watcher(req, res, next) {
-    console.log("/")
-    if (!req.session.username) res.redirect("/signIn");
-    else next();
-  }
-
   #routes() {
-    this.router.get("/", this.#session_watcher, ctrlUser.index);
-    this.router.get("/signUp/", ctrlUser.viewSignUp);
-    this.router.get("/signIn/", ctrlUser.viewSignIn);
     this.router.post("/signIn", ctrlUser.signIn);
     this.router.get("/log_out", ctrlUser.logOut);
     this.router.get("/users/", ctrlUser.getUsers);
